+++ packages/perseus/src/widget-ai-utils/image/image-ai-utils.test.ts
@@ -1,5 +1,7 @@
 import {renderQuestion} from "../../widgets/__testutils__/renderQuestion";
 
+import {getPromptJSON} from "./image-ai-utils";
+
 import type {ImageWidget} from "../../perseus-types";
 
 const question = {
@@ -37,7 +39,30 @@ const question = {
     },
 } as const;
 
-describe("image widget", () => {
+describe("Image AI utils", () => {
+    it("it returns JSON with the expected format and fields", () => {
+        const renderProps: any = {
+            alt: "An image of a textbook",
+            title: "Textbook",
+            caption: "A textbook",
+            backgroundImage: {
+                url: "https://www.khanacademy.org/some-image.png",
+            },
+        };
+
+        const resultJSON = getPromptJSON(renderProps);
+
+        expect(resultJSON).toEqual({
+            type: "image",
+            options: {
+                altText: "An image of a textbook",
+                title: "Textbook",
+                caption: "A textbook",
+                imageUrl: "https://www.khanacademy.org/some-image.png",
+            },
+        });
+    });
+
     it("should get prompt json which matches the state of the UI", async () => {
         // Arrange
         const {renderer} = renderQuestion(question, {isMobile: false});
