+++ packages/perseus/src/widget-ai-utils/image/image-ai-utils.ts
