--- packages/perseus/src/widget-ai-utils/input-number/prompt-utils.test.ts
@@ -1,29 +1,0 @@
-import {getPromptJSON} from "./prompt-utils";
-
-import type {PerseusInputNumberUserInput} from "../../validation.types";
-
-describe("InputNumber getPromptJSON", () => {
-    it("it returns JSON with the expected format and fields", () => {
-        const renderProps: any = {
-            simplify: "optional",
-            answerType: "integer",
-        };
-
-        const userInput: PerseusInputNumberUserInput = {
-            currentValue: "123",
-        };
-
-        const resultJSON = getPromptJSON(renderProps, userInput);
-
-        expect(resultJSON).toEqual({
-            type: "input-number",
-            options: {
-                simplify: "optional",
-                answerType: "integer",
-            },
-            userInput: {
-                value: "123",
-            },
-        });
-    });
-});
